Move functions into module with independence
